# Optimize `iszero` function (3-5x faster)