--- a/datafusion/functions/Cargo.toml
+++ b/datafusion/functions/Cargo.toml
@@ -119,6 +119,11 @@ required-features = ["datetime_expressions"]
 
 [[bench]]
 harness = false
+name = "iszero"
+required-features = ["math_expressions"]
+
+[[bench]]
+harness = false
 name = "nullif"
 required-features = ["core_expressions"]